Update to InfectedHordes
Can turn on and off see through zombies via the json. Variable name
You can now toggle the horde spawn message via the json. Regardless, there is log messages inside of script.log

Setting the lifetime to 0 no longer deletes the hordes when they spawn, instead when all infected from the  horde are killed off they are removed from the hordes, and the new ones can spawn, if 1 exists then the horde still lives on. This allows horde to live longer on the map, and allow them to travel and spread out. The Infected bodies should be removed with normal removal.

Check to make sure locations are not empty.  Check if random location is free of players before spawning the horde. It will only check random locations hordePosition.Count * 2, so if by chance it loops 20 times, and never finds a point it will return null and not spawn a horde.